# src/config/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MateTo API",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/auth/signup": {
      "post": {
        "summary": "SignUp",
        "x-apidog-folder": "Authentication",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/413215/apis/api-5093886-run"
      }
    },
    "/auth/signin": {
      "post": {
        "summary": "SignIn",
        "x-apidog-folder": "Authentication",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Authentication"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/413215/apis/api-5093923-run"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Development server"
    }
  ]
}